+++ crates/builder/validated-tx-rpc/Cargo.toml
@@ -13,27 +13,19 @@ workspace = true
 
 [dependencies]
 # workspace
+base-primitives.workspace = true
 base-client-node.workspace = true
 base-alloy-consensus.workspace = true
 
 # reth
 reth-transaction-pool.workspace = true
-reth-primitives-traits.workspace = true
 reth-optimism-txpool.workspace = true
-
-# alloy
-alloy-primitives.workspace = true
-alloy-consensus.workspace = true
-alloy-eips.workspace = true
 
 # rpc
 jsonrpsee.workspace = true
 
 # misc
-serde.workspace = true
 tracing.workspace = true
-thiserror.workspace = true
-base64.workspace = true
 
 [dev-dependencies]
 eyre.workspace = true
