+++ crates/builder/validated-tx-rpc/src/lib.rs
@@ -3,11 +3,10 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
-mod convert;
-pub use convert::{ConversionError, compute_tx_hash};
-
-mod types;
-pub use types::{InsertResult, TransactionSignature, ValidatedTransaction, base64_bytes};
+// Re-export types from shared primitives
+pub use base_primitives::{
+    ConversionError, InsertResult, TransactionSignature, ValidatedTransaction, base64_bytes,
+};
 
 mod rpc;
 pub use rpc::{ValidatedTxApiImpl, ValidatedTxApiServer};
